pyproject.toml
[project]
name = "morning-call"
version = "0.1.0"
description = "20-minute morning audio news briefing"
requires-python = ">=3.9"
dependencies = [
    "feedparser>=6.0",
    "httpx>=0.27",
    "edge-tts>=6.1",
    "pydub>=0.25",
    "readability-lxml>=0.8",
]

[project.scripts]
morning-call = "morning_call.main:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
